+++ example/android/app/src/main/kotlin/com/dcc/file_previewer/file_previewer_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.dcc.file_previewer.file_previewer_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}