+++ ios/Classes/FilePreviewPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FilePreviewPlugin : NSObject<FlutterPlugin>
-@end